Bug Fixed: href links
Solution: the transition I was using with the observer tool was translating in the y-direction. While it looks cool, this caused a problem where with each load, there'd be extra space in the y-direction which messed with the way the href worked. Changed to x-translation and it works fine. Looks just as fine too.